+++ asset_management/app/Http/Controllers/SupplierController.php
@@ -34,6 +34,7 @@ class SupplierController extends Controller
 
     public function store(Request $request)
     {
+        $this->authorize('create-delete-users');
         $request->validate([
             'name' => 'required|string|max:255|unique:suppliers',
         ]);
@@ -47,10 +48,25 @@ class SupplierController extends Controller
 
     public function destroy(Supplier $supplier)
     {
+        $this->authorize('create-delete-users');
         // Implement logic to delete the supplier
         // Don't forget to handle any related dependencies, if any
         $supplier->delete();
 
         return response()->json(null, 204);
     }
+
+    public function update(Request $request, Supplier $sup)
+    {
+        $this->authorize('create-delete-users');
+        $request->validate([
+            'name' => 'required|string|max:255' . $sup->id,
+        ]);
+
+        $sup->update([
+            'name' => $request->input('name'),
+        ]);
+
+        return response()->json($sup, 200);
+    } 
 }
